requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-dotenv>=1.0.0
